% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.stable_population_free_fit}
\alias{print.stable_population_free_fit}
\title{Print a free Weibull fit}
\usage{
\method{print}{stable_population_free_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"stable_population_free_fit"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a free Weibull fit
}
