% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.stable_population_reconstruction}
\alias{print.stable_population_reconstruction}
\title{Print a stable population reconstruction}
\usage{
\method{print}{stable_population_reconstruction}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"stable_population_reconstruction"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a stable population reconstruction
}
