% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_alpha.R
\name{alpha_objective}
\alias{alpha_objective}
\title{Objective function for \code{uniroot}: finds the difference between births and 1}
\usage{
alpha_objective(alpha, beta, fertility_rates)
}
\arguments{
\item{alpha}{A numeric value representing the alpha parameter.}

\item{beta}{A numeric value representing the beta parameter.}

\item{fertility_rates}{A numeric vector containing the fertility rates.}
}
\value{
The difference between the number of births and 1.
}
\description{
This function calculates the difference between the number of births calculated with the given values of alpha, beta, and fertility rates, and the target value of 1.
}
\examples{
alpha_objective(0.5, 1.2, c(0.2, 0.3, 0.5, 0.4))
}
