% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{calculate_population}
\alias{calculate_population}
\title{Calculates the population for each age group}
\usage{
calculate_population(alpha, beta, fertility_rates)
}
\arguments{
\item{alpha}{A numeric value representing the population growth rate.}

\item{beta}{A numeric value affecting the birth rate.}

\item{fertility_rates}{A vector of fertility rates for each age group.}
}
\value{
A list containing the calculated population and births.
}
\description{
This function calculates the population for each age group and the number of births.
}
\examples{
calculate_population(0.5, 1.2, c(0.2, 0.3, 0.5, 0.4))
}
