% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{find_stablepopulations_root}
\alias{find_stablepopulations_root}
\title{Find Root Directory of StablePopulation}
\usage{
find_stablepopulations_root()
}
\value{
The full path to the \code{StablePopulation} directory if found.
}
\description{
This internal function searches for the root directory of the \code{StablePopulation}
project by looking for a folder named \code{StablePopulation} in the current
or parent directories. It is used internally to locate project-specific files.
}
\keyword{internal}
