% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_analysis}
\alias{run_analysis}
\title{Run Analysis on Excel Data and Export Results}
\usage{
run_analysis()
}
\description{
This function reads fertility rate data and Beta value from an Excel file, processes it, and exports
the results to a new Excel file for the specie, including population matrices and calculated alpha/beta values.
}
