% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{weibull_survival}
\alias{weibull_survival}
\title{Weibull function for the survival rate}
\usage{
weibull_survival(alpha, beta, age)
}
\arguments{
\item{alpha}{A numeric value representing the scale parameter of the Weibull distribution.}

\item{beta}{A numeric value representing the shape parameter of the Weibull distribution.}

\item{age}{A numeric value representing the age.}
}
\value{
A numeric value giving the survival rate (probability) for reaching the given age.
}
\description{
This function calculates the survival rate to reach a specific age using the Weibull function.
}
\examples{
weibull_survival(1.5, 0.8, 10)
}
